Extract file subcommand to cmd_file.go
Move file flags, command definition, and run function to separate file.
